Implement optimistic rendering for UPDATE_MANY
