+++ packages/ra-core/src/sideEffect/saga/crudSaga.js
@@ -2,6 +2,7 @@ import { all } from 'redux-saga/effects';
 import auth from './auth';
 import crudFetch from './crudFetch';
 import crudResponse from './crudResponse';
+import optimistic from './optimistic';
 import referenceFetch from './referenceFetch';
 import i18n from './i18n';
 
@@ -16,5 +17,6 @@ export default (dataProvider, authProvider, i18nProvider) =>
             crudFetch(dataProvider)(),
             crudResponse(),
             referenceFetch(),
+            optimistic(),
         ]);
     };
