+++ packages/ra-core/src/sideEffect/saga/index.js
@@ -2,5 +2,6 @@ export auth from './auth';
 export crudFetch from './crudFetch';
 export crudResponse from './crudResponse';
 export crudSaga from './crudSaga';
+export optimistic from './optimistic';
 export referenceFetch from './referenceFetch';
 export i18n from './i18n';
