+++ packages/ra-ui-materialui/src/list/DatagridBody.js
@@ -21,6 +21,7 @@ const DatagridBody = ({
     styles,
     rowStyle,
     onToggleItem,
+    version,
     ...rest
 }) => (
     <TableBody className={classnames('datagrid-body', className)} {...rest}>
@@ -71,13 +72,9 @@ DatagridBody.defaultProps = {
     ids: [],
 };
 
-const areArraysEqual = (arr1, arr2) =>
-    arr1.length == arr2.length && arr1.every((v, i) => v === arr2[i]);
-
 const PureDatagridBody = shouldUpdate(
     (props, nextProps) =>
-        !areArraysEqual(props.ids, nextProps.ids) ||
-        nextProps.isLoading === false
+        props.version !== nextProps.version || nextProps.isLoading === false
 )(DatagridBody);
 
 // trick material-ui Table into thinking this is one of the child type it supports
