+++ packages/ra-ui-materialui/src/list/List.js
@@ -150,6 +150,7 @@ export const ListView = ({
                                 resource,
                                 selectedIds,
                                 setSort,
+                                version,
                             })}
                         {!isLoading &&
                             !ids.length && (
